+++ docs/operations.rst
@@ -1,12 +1,12 @@
 .. contents::
 
-.. _custom_ops:
+.. _operations:
 
 **********
-custom_ops
+operations
 **********
 This module consists of a pre-defined functions that are imported into
 :ref:`tag` for convenient use in ``VirtualTag`` objects.
 
-.. automodule:: pype_schema.custom_ops
+.. automodule:: pype_schema.operations
    :members:
